chore: move model options dialog in separate folder
